+++ generator/src/main/java/org/ehrbase/client/classgenerator/FlattFilter.java
@@ -24,11 +24,13 @@ import com.nedap.archie.rminfo.RMTypeInfo;
 import org.apache.commons.collections4.SetUtils;
 import org.ehrbase.serialisation.util.SnakeCase;
 import org.ehrbase.util.reflection.ReflectionHelper;
+import org.ehrbase.util.rmconstants.RmConstants;
 import org.ehrbase.webtemplate.filter.Filter;
 import org.ehrbase.webtemplate.model.WebTemplate;
 import org.ehrbase.webtemplate.model.WebTemplateNode;
 import org.ehrbase.webtemplate.parser.FlatPath;
 import org.ehrbase.webtemplate.parser.config.RmIntrospectConfig;
+import org.ehrbase.webtemplate.util.WebTemplateUtils;
 
 import java.util.*;
 import java.util.stream.Collectors;
@@ -98,16 +100,19 @@ public class FlattFilter extends Filter {
         return !node.getChildren().isEmpty()
             && node.getMax() == 1
             && !node.getRmType().equals("COMPOSITION")
+            && isSkippableInterval(parent, node)
             && (!isEvent(node) || isSkippableEvent(parent, node));
       case SECTION:
         return !node.getChildren().isEmpty()
             && node.getMax() == 1
             && (!node.isArchetype() || node.getRmType().equals("SECTION"))
+            && isSkippableInterval(parent, node)
             && (!isEvent(node) || isSkippableEvent(parent, node));
       default:
         return !node.getChildren().isEmpty()
             && node.getMax() == 1
             && !node.isArchetype()
+            && isSkippableInterval(parent, node)
             && (!isEvent(node) || isSkippableEvent(parent, node));
     }
   }
@@ -118,6 +123,15 @@ public class FlattFilter extends Filter {
       return false;
     }
     return parent.getChildren().stream().filter(this::isEvent).count() == 1 && !node.isMulti();
+  }
+
+  private boolean isSkippableInterval(WebTemplateNode parent, WebTemplateNode node) {
+
+    if (parent.getRmType().equals(RmConstants.ELEMENT)
+        && node.getRmType().contains("DV_INTERVAL")) {
+      return WebTemplateUtils.getTrueChildrenElement(parent).size() == 1;
+    }
+    return true;
   }
 
   protected void preHandle(WebTemplateNode node) {