+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Note: version releases in the 0.x.y range may introduce breaking changes.
 - Validation: fix validation of DV_TEXT with listopen (see https://github.com/ehrbase/openEHR_SDK/pull/335) 
 - Walker, Flat, DTO, Validation  : enhance performance of path handling (
   see https://github.com/ehrbase/openEHR_SDK/pull/325 )
+- dto: Fix handling of element wich contains a choice with one an interval (see https://github.com/ehrbase/openEHR_SDK/pull/334)
 
 ## 1.17.0